requirements.txt
datasets>=2.0.0
huggingface_hub>=0.18.0
sentence-transformers>=2.2.2
transformers>=4.34.0
torch>=2.0.0
faiss-cpu>=1.7.4
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pytest>=7.4.0
numpy>=1.24.0
pydantic>=2.4.2
